converint to float instead of an int

+++ Testing/k0/aggregate_data_k0_rmse.py
@@ -46,7 +46,7 @@ for line in Lines:
     if 'Test - k0' in line:
         k0_val = int(re.findall('-?\d+', line)[1])
     if 'RMSE' in line:
-        rmse = list(map(int, re.findall("\d+\.\d+", line)))[0]
+        rmse = list(map(float, re.findall("\d+\.\d+", line)))[0]
         i = k0.index(k0_val)
         final_matrices[current_data_type]['rmse'][i] = rmse
